+++ typescript/cli/src/tests/warp/warp-check.e2e-test.ts
@@ -17,13 +17,10 @@ import {
   CHAIN_NAME_3,
   CORE_CONFIG_PATH,
   DEFAULT_E2E_TEST_TIMEOUT,
-  KeyBoardKeys,
-  TestPromptAction,
   WARP_DEPLOY_OUTPUT_PATH,
   deployOrUseExistingCore,
   deployToken,
   getCombinedWarpRoutePath,
-  handlePrompts,
 } from '../commands/helpers.js';
 import {
   hyperlaneWarpCheck,
@@ -97,30 +94,28 @@ describe('hyperlane warp check e2e tests', async function () {
     return warpReadResult;
   }
 
-  describe('HYP_KEY=... hyperlane warp check --config ...', () => {
+  describe('hyperlane warp check --config ...', () => {
     it(`should require both warp core & warp deploy config paths to be provided together`, async function () {
       await deployAndExportWarpRoute(tokenSymbol, token.address);
 
-      const finalOutput = await hyperlaneWarpCheckRaw({
-        hypKey: ANVIL_KEY,
+      const output = await hyperlaneWarpCheckRaw({
         warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,
       })
         .stdio('pipe')
         .nothrow();
 
-      expect(finalOutput.exitCode).to.equal(1);
-      expect(finalOutput.text()).to.include(
+      expect(output.exitCode).to.equal(1);
+      expect(output.text()).to.include(
         'Both --config/-i and --warp/-wc must be provided together when using individual file paths',
       );
     });
   });
 
-  describe('hyperlane warp check --key ... --config ...', () => {
+  describe('hyperlane warp check --config ...', () => {
     it(`should require both warp core & warp deploy config paths to be provided together`, async function () {
       await deployAndExportWarpRoute(tokenSymbol, token.address);
 
       const finalOutput = await hyperlaneWarpCheckRaw({
-        privateKey: ANVIL_KEY,
         warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,
       })
         .stdio('pipe')
@@ -137,19 +132,6 @@ describe('hyperlane warp check e2e tests', async function () {
     it(`should not find any differences between the on chain config and the local one`, async function () {
       await deployAndExportWarpRoute(tokenSymbol, token.address);
 
-      const steps: TestPromptAction[] = [
-        {
-          check: (currentOutput) =>
-            currentOutput.includes('Please enter the private key for chain'),
-          input: `${ANVIL_KEY}${KeyBoardKeys.ENTER}`,
-        },
-        {
-          check: (currentOutput) =>
-            currentOutput.includes('Please enter the private key for chain'),
-          input: `${ANVIL_KEY}${KeyBoardKeys.ENTER}`,
-        },
-      ];
-
       const output = hyperlaneWarpCheckRaw({
         symbol: tokenSymbol,
         warpDeployPath: WARP_DEPLOY_OUTPUT_PATH,
@@ -158,10 +140,8 @@ describe('hyperlane warp check e2e tests', async function () {
         .stdio('pipe')
         .nothrow();
 
-      const finalOutput = await handlePrompts(output, steps);
-
-      expect(finalOutput.exitCode).to.equal(0);
-      expect(finalOutput.text()).to.include('No violations found');
+      expect(output.exitCode).to.equal(0);
+      expect(output.text()).to.include('No violations found');
     });
   });