+++ typescript/infra/config/environments/testnet4/funding.ts
@@ -10,7 +10,7 @@ export const keyFunderConfig: KeyFunderConfig<
 > = {
   docker: {
     repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-    tag: 'b4b4bf5-20250207-143914',
+    tag: 'fb4d433-20250303-180119',
   },
   // We're currently using the same deployer key as testnet2.
   // To minimize nonce clobbering we offset the key funder cron
@@ -47,6 +47,7 @@ export const keyFunderConfig: KeyFunderConfig<
     holesky: '5',
     hyperliquidevmtestnet: '0.1',
     inksepolia: '0.1',
+    monadtestnet: '0.1',
     odysseytestnet: '0.1',
     optimismsepolia: '0.1',
     // Disabling plumetestnet on Sept 16, 2024: chain is paused for "airplane mode"
@@ -65,6 +66,7 @@ export const keyFunderConfig: KeyFunderConfig<
     superpositiontestnet: '1',
     treasuretopaz: '5',
     unichaintestnet: '0.1',
+    weavevmtestnet: '0.1',
   },
   desiredKathyBalancePerChain: {
     alfajores: '1',
@@ -90,6 +92,8 @@ export const keyFunderConfig: KeyFunderConfig<
   igpClaimThresholdPerChain: {
     alfajores: '1',
     arbitrumsepolia: '0.05',
+    // never claim on berachain berabartio testnet
+    berabartio: '1000.0',
     basesepolia: '0.05',
     bsctestnet: '1',
     connextsepolia: '0.1',