+++ typescript/infra/config/environments/testnet4/supportedChainNames.ts
@@ -20,6 +20,7 @@ export const testnet4SupportedChainNames = [
   'holesky',
   'hyperliquidevmtestnet',
   'inksepolia',
+  'monadtestnet',
   'odysseytestnet',
   'optimismsepolia',
   // Disabling plumetestnet on Sept 16, 2024: chain is paused for "airplane mode"
@@ -36,6 +37,7 @@ export const testnet4SupportedChainNames = [
   'superpositiontestnet',
   'treasuretopaz',
   'unichaintestnet',
+  'weavevmtestnet',
 ] as const;
 
 export const supportedChainNames = [...testnet4SupportedChainNames];