+++ typescript/infra/scripts/funding/write-alert.ts
@@ -1,42 +1,48 @@
-import { checkbox } from '@inquirer/prompts';
-import yargs from 'yargs';
-
 import { ChainMap } from '@hyperlane-xyz/sdk';
 import { rootLogger } from '@hyperlane-xyz/utils';
 
-import { THRESHOLD_CONFIG_PATH } from '../../src/config/funding/balances.js';
+import {
+  BalanceThresholdType,
+  THRESHOLD_CONFIG_PATH,
+  ThresholdsData,
+  balanceThresholdConfigMapping,
+} from '../../src/config/funding/balances.js';
 import {
   AlertType,
   alertConfigMapping,
 } from '../../src/config/funding/grafanaAlerts.js';
+import { validateThresholds } from '../../src/funding/balances.js';
 import {
   fetchGrafanaAlert,
-  fetchServiceAccountToken,
+  fetchGrafanaServiceAccountToken,
   generateQuery,
   updateGrafanaAlert,
-} from '../../src/funding/grafana.js';
+} from '../../src/infrastructure/monitoring/grafana.js';
 import { readJSONAtPath } from '../../src/utils/utils.js';
-import { withAlertType, withConfirmAllChoices } from '../agent-utils.js';
 
 async function main() {
-  const { alertType, all } = await withConfirmAllChoices(
-    withAlertType(yargs(process.argv.slice(2))),
-  ).argv;
+  const saToken = await fetchGrafanaServiceAccountToken();
 
-  const saToken = await fetchServiceAccountToken();
+  const balanceThresholdTypes = Object.values(BalanceThresholdType);
+  const balanceThresholdConfigs: ThresholdsData = balanceThresholdTypes.reduce(
+    (acc, balanceThresholdType) => {
+      const thresholds = readJSONAtPath(
+        `${THRESHOLD_CONFIG_PATH}/${balanceThresholdConfigMapping[balanceThresholdType].configFileName}`,
+      ) as ChainMap<string>;
 
-  const alertsToUpdate: AlertType[] = all
-    ? Object.values(AlertType)
-    : alertType
-    ? [alertType]
-    : await checkbox({
-        message: 'Select the alert type to update',
-        choices: Object.values(AlertType).map((alert) => ({
-          name: alertConfigMapping[alert].choiceLabel,
-          value: alert,
-          checked: true, // default to all checked
-        })),
-      });
+      return {
+        ...acc,
+        [balanceThresholdType]: {
+          thresholds,
+        },
+      };
+    },
+    {} as ThresholdsData,
+  );
+
+  validateThresholds(balanceThresholdConfigs);
+
+  const alertsToUpdate = Object.values(AlertType);
 
   for (const alert of alertsToUpdate) {
     // fetch alertRule config from Grafana